+++ src/render/OutlineMap.js
@@ -1,13 +1,13 @@
 
-render.AmbientMap = {
+render.OutlineMap = {
 
   init: function() {
     this.shader = new glx.Shader({
-      vertexShader:   Shaders.ambientFromDepth.vertex,
-      fragmentShader: Shaders.ambientFromDepth.fragment,
-      shaderName: 'SSAO shader',
+      vertexShader:   Shaders.outlineMap.vertex,
+      fragmentShader: Shaders.outlineMap.fragment,
+      shaderName: 'outline map shader',
       attributes: ['aPosition', 'aTexCoord'],
-      uniforms: ['uInverseTexSize', 'uNearPlane', 'uFarPlane', 'uDepthTexIndex', 'uFogTexIndex', 'uEffectStrength']
+      uniforms: ['uMatrix', 'uInverseTexSize', 'uNearPlane', 'uFarPlane', 'uDepthTexIndex', 'uFogNormalTexIndex', 'uIdTexIndex', 'uEffectStrength']
     });
 
     this.framebuffer = new glx.Framebuffer(128, 128); //dummy value, size will be set dynamically
@@ -31,7 +31,7 @@ render.AmbientMap = {
     ]));
   },
 
-  render: function(depthTexture, fogTexture, framebufferSize, effectStrength) {
+  render: function(depthTexture, fogNormalTexture, idTexture, framebufferSize, effectStrength) {
 
     var
       shader = this.shader,
@@ -50,18 +50,21 @@ render.AmbientMap = {
     gl.clearColor(1.0, 0.0, 0.0, 1);
     gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
 
+    gl.uniformMatrix4fv(shader.uniforms.uMatrix, false, glx.Matrix.identity().data);
+
     shader.setUniforms([
       ['uInverseTexSize', '2fv', [1/framebufferSize[0], 1/framebufferSize[1]]],
       ['uEffectStrength', '1f',  effectStrength],
       ['uNearPlane',      '1f',  1.0], //FIXME: use actual near and far planes of the projection matrix
-      ['uFarPlane',       '1f',  7500.0]
+      ['uFarPlane',       '1f',  7500.0]      
     ]);
 
     shader.bindBuffer(this.vertexBuffer,   'aPosition');
     shader.bindBuffer(this.texCoordBuffer, 'aTexCoord');
 
-    shader.bindTexture('uDepthTexIndex', 0, depthTexture);
-    shader.bindTexture('uFogTexIndex',   1, fogTexture);
+    shader.bindTexture('uDepthTexIndex',    0, depthTexture);
+    shader.bindTexture('uFogNormalTexIndex',1, fogNormalTexture);
+    shader.bindTexture('uIdTexIndex',       2, idTexture);
 
     gl.drawArrays(gl.TRIANGLES, 0, this.vertexBuffer.numItems);